added crud for farmer marketplace

+++ model/siklus.go
@@ -6,25 +6,25 @@ import (
 )
 
 type Siklus struct {
-	ID            uint   `gorm:"primarykey" json:"id"`
-	Tanggal       string `json:"tanggal"`
-	DocSiklus     int    `json:"doc_siklus"`
-	PHRealtime    int    `json:"ph_realtime"`
-	PHPagi        int    `json:"ph_pagi"`
-	PHSiang       int    `json:"ph_siang"`
-	PHMalam       int    `json:"ph_malam"`
-	SuhuRealtime  int    `json:"suhu_realtime"`
-	SuhuPagi      int    `json:"suhu_pagi"`
-	SuhuSiang     int    `json:"suhu_siang"`
-	SuhuMalam     int    `json:"suhu_malam"`
-	DORealtime    int    `json:"do_realtime"`
-	DOPagi        int    `json:"do_pagi"`
-	DOSiang       int    `json:"do_siang"`
-	DOMalam       int    `json:"do_malam"`
-	GaramRealtime int    `json:"garam_realtime"`
-	GaramPagi     int    `json:"garam_pagi"`
-	GaramSiang    int    `json:"garam_siang"`
-	GaramMalam    int    `json:"garam_malam"`
+	ID            uint    `gorm:"primarykey" json:"id"`
+	Tanggal       string  `json:"tanggal"`
+	DocSiklus     int     `json:"doc_siklus"`
+	PHRealtime    float64 `json:"ph_realtime"`
+	PHPagi        float64 `json:"ph_pagi"`
+	PHSiang       float64 `json:"ph_siang"`
+	PHMalam       float64 `json:"ph_malam"`
+	SuhuRealtime  float64 `json:"suhu_realtime"`
+	SuhuPagi      float64 `json:"suhu_pagi"`
+	SuhuSiang     float64 `json:"suhu_siang"`
+	SuhuMalam     float64 `json:"suhu_malam"`
+	DORealtime    float64 `json:"do_realtime"`
+	DOPagi        float64 `json:"do_pagi"`
+	DOSiang       float64 `json:"do_siang"`
+	DOMalam       float64 `json:"do_malam"`
+	GaramRealtime float64 `json:"garam_realtime"`
+	GaramPagi     float64 `json:"garam_pagi"`
+	GaramSiang    float64 `json:"garam_siang"`
+	GaramMalam    float64 `json:"garam_malam"`
 
 	AquaFarmer   AquaFarmer `gorm:"ForeignKey:AquaFarmerID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
 	AquaFarmerID uuid.UUID  `gorm:"null" json:"user_id"`
